Add hash_utils
